+++ src/Intercom/Data/Segment.cs
@@ -13,8 +13,8 @@ namespace Intercom.Data
 	public class Segment : Model
 	{
         public string name { get; set; }
-        public int created_at { get; set; }
-        public int updated_at { get; set; }
+        public long created_at { get; set; }
+        public long updated_at { get; set; }
 
 		public Segment ()
 		{
